server/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==4.2.17
django-allauth==65.3.0
django-cors-headers==4.4.0
dj-rest-auth==7.0.0
django-debug-toolbar==4.4.0
djangorestframework==3.15.0
django-extensions==3.0.8
django-anymail[mailjet]==12.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.25.3
wrapt==1.11.2
psycopg==3.1.8
whitenoise==5.2.0
newrelic==10.7.0
django-cockroachdb==4.2.*
